requirements.txt
### dependency
tqdm
pandas
numpy
PyYAML
flask
shutils
black # for simple code formatting
sklearn-pandas
matplotlib
kneed
# data and model versioning
dvc
mlflow
# packages for model training
xgboost
scikit-learn
### local packages -
-e . # << uncomment this to install from setup.py
